+++ component/include/pl_esp_ethernet.h
@@ -46,6 +46,7 @@ public:
 private:
   Mutex mutex;
   esp_netif_t* netif = NULL;
+  esp_event_handler_instance_t eventHandlerInstance = NULL;
   bool enabled = false, connected = false;
   PhyNewFunction phyNewFunction;
   eth_phy_config_t phyConfig = ETH_PHY_DEFAULT_CONFIG();